api/pyproject.toml
[project]
name = "detonate"
version = "0.1.0"
description = "Open-source malware analysis sandbox"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "minio>=7.2.0",
    "python-magic>=0.4.27",
    "redis>=5.0.0",
    "celery>=5.4.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "anthropic>=0.40.0",
    "ollama>=0.4.0",
    "pefile>=2024.8.26",
    "sse-starlette>=2.0.0",
    "docker>=7.0.0",
    "yara-python>=4.5.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "websockify>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
